tests/wpt/tests/html/browsers/history/the-location-interface/location-valueof.html

<!DOCTYPE html>
<title>Location valueOf</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id=log></div>
<script>
test(() => {
  assert_equals(location.valueOf, Object.prototype.valueOf)
  assert_equals(typeof location.valueOf.call(5), "object")
  const prop = Object.getOwnPropertyDescriptor(location, "valueOf")
  assert_false(prop.enumerable)
  assert_false(prop.writable)
  assert_false(prop.configurable)
})
</script>
