third_party/WebKit/LayoutTests/external/wpt/html/browsers/history/the-location-interface/location-valueof-expected.txt

This is a testharness.js-based test.
FAIL Location valueOf assert_equals: expected function "function valueOf() { [native code] }" but got function "function valueOf() { [native code] }"
Harness: the test ran to completion.
